cdk_container_deploy_app/requirements.txt
aws-cdk-lib==2.169.0
constructs>=10.0.0,<11.0.0
# aws-cdk.aws-apprunner-alpha
# aws-cdk.aws-ecr-assets
requests
